Modularize the render functions into renderBoard and renderCell
